requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.1,<3.0
djangorestframework
drf_yasg

